+++ knytt/objects/banks/bank12/GhostMarker.cs
@@ -1,0 +1,16 @@
+using Godot;
+
+public class GhostMarker : GhostObject
+{
+    protected override void _InvReady()
+    {
+        flickerMax = .2f;
+        flickerMin = .8f;
+        flip_time = .03f;
+        change_fraction = 16f;
+
+        GetNode<AnimatedSprite>("AnimatedSprite").Frame = ObjectID.y - 18;
+    }
+
+    protected override void _impl_initialize() { }
+}